fix(imageviewer): perform window sizing in float as much as possible

+++ include/tev/ImageViewer.h
@@ -148,9 +148,9 @@ public:
     EDisplayWhiteLevelSetting displayWhiteLevelSetting() const;
     void setDisplayWhiteLevelSetting(EDisplayWhiteLevelSetting value);
 
-    nanogui::Vector2i sizeToFitImage(const std::shared_ptr<Image>& image);
-    nanogui::Vector2i sizeToFitAllImages();
-    void resizeToFit(nanogui::Vector2i size);
+    nanogui::Vector2f sizeToFitImage(const std::shared_ptr<Image>& image);
+    nanogui::Vector2f sizeToFitAllImages();
+    void resizeToFit(nanogui::Vector2f size);
 
     bool playingBack() const;
     void setPlayingBack(bool value);
@@ -355,8 +355,8 @@ private:
     // Misc state tracking variables
     int mDidFitToImage = 0;
 
-    nanogui::Vector2i mMinWindowPos = {0, 0};
-    nanogui::Vector2i mMaxWindowSize = {8192, 8192};
+    nanogui::Vector2f mMinWindowPos = {0, 0};
+    nanogui::Vector2f mMaxWindowSize = {8192, 8192};
 
     bool mInitialized = false;
